// src/module_bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const RepairOrder = __t.object("RepairOrder", {
  number: __t.u32(),
  customer: __t.option(__t.string()),
  allPartsAreHere: __t.bool(),
  transferOrders: __t.array(__t.u32()),
  get techName() {
    return TechName;
  },
  notes: __t.string(),
});
export type RepairOrder = __Infer<typeof RepairOrder>;

// The tagged union or sum type for the algebraic type `TechName`.
export const TechName = __t.enum("TechName", {
  Aaron: __t.unit(),
  Andrew: __t.unit(),
  Chris: __t.unit(),
  Jeremiah: __t.unit(),
  Jr: __t.unit(),
  Justin: __t.unit(),
  Nick: __t.unit(),
  Other: __t.unit(),
});
export type TechName = __Infer<typeof TechName>;